cmd/geth, cmd/utils, ethdb/remotedb: refactor a bit
